+++ extensions/backstage/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Backstage Changelog
 
+## [Support static token] - {PR_MERGE_DATE}
+
+- address [issue 20382](https://github.com/raycast/extensions/issues/20382) and
+  support Backstage static token authentication
+
 ## [Initial Release] - 2024-08-08
 
 - the initial release of Backstage extension, including Google Cloud IAP support.